refactor: separate interfaces (DD-3)
