+++ src/app/features/auth/interfaces/user-request.interface.ts
@@ -4,7 +4,3 @@ export interface UserRequest {
   email: string;
   password: string;
 }
-
-export interface UserResponse {
-  token: string;
-}
