+++ src/app/features/auth/interfaces/user-response.interface.ts
@@ -1,0 +1,3 @@
+export interface UserResponse {
+  token: string;
+}
